wire up a Si7021 I2C temperature and humidity sensor

+++ src/HomeMonitor.Api/Temperature/GraphQL/TemperatureMeasurement.cs
@@ -6,7 +6,7 @@ public class TemperatureMeasurement : ITemperatureMeasured
 {
     public Guid Id { get; init; }
     public string Location { get; init; }
-    public decimal Temperature { get; init; }
-    public decimal Humidity { get; init; }
+    public double Temperature { get; init; }
+    public double Humidity { get; init; }
     public DateTimeOffset RecordedAt { get; init; }
 }
